+++ changelog.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+#### 20160914.1
+* Modified: Add title text, remove logo image
+* Updated: added standard changes that I use on a child theme.
+
 #### 20160512.1
 * Updated: theme description, bumped version